refactor: Apps was converted to separate pages and redesigned for vue 3
